+++ Dependencies/MelonStartScreen/UI/UICustomization.cs
@@ -54,6 +54,7 @@ namespace MelonLoader.MelonStartScreen
             internal Color SolidColor = new Color(0.08f, 0.09f, 0.10f);
             [TomlPrecedingComment("If should Scan for and Load Custom Background Image  ( true | false )")]
             internal bool ScanForCustomImage = true;
+
             [TomlPrecedingComment("UnityEngine.FilterMode of the Background Image  ( \"Point\" | \"Bilinear\" | \"Trilinear\" )")]
             internal FilterMode Filter = FilterMode.Bilinear;
         }
@@ -64,9 +65,11 @@ namespace MelonLoader.MelonStartScreen
             internal bool Enabled = true;
             [TomlPrecedingComment("If should Scan for and Load Custom Logo Image  ( true | false )")]
             internal bool ScanForCustomImage = true;
+
             [TomlPrecedingComment("UnityEngine.FilterMode of the Logo Image  ( \"Point\" | \"Bilinear\" | \"Trilinear\" )")]
             internal FilterMode Filter = FilterMode.Bilinear;
-            [TomlPrecedingComment("Toggles the Auto-Alignment of the Logo Image")]
+
+            [TomlPrecedingComment("Toggles the Position Auto-Alignment of the Logo Image")]
             internal bool AutoAlign = true;
             [TomlPrecedingComment("Custom Position of the Logo Image")]
             internal LemonTuple<int, int> CustomPosition = new LemonTuple<int, int>();
@@ -78,9 +81,11 @@ namespace MelonLoader.MelonStartScreen
             internal bool Enabled = true;
             [TomlPrecedingComment("If should Scan for and Load Custom Loading Image  ( true | false )")]
             internal bool ScanForCustomImage = true;
+
             [TomlPrecedingComment("UnityEngine.FilterMode of the Loading Image  ( \"Point\" | \"Bilinear\" | \"Trilinear\" )")]
             internal FilterMode Filter = FilterMode.Bilinear;
-            [TomlPrecedingComment("Toggles the Auto-Alignment of the Loading Image")]
+
+            [TomlPrecedingComment("Toggles the Position Auto-Alignment of the Loading Image")]
             internal bool AutoAlign = true;
             [TomlPrecedingComment("Custom Position of the Loading Image")]
             internal LemonTuple<int, int> CustomPosition = new LemonTuple<int, int>();
@@ -92,10 +97,20 @@ namespace MelonLoader.MelonStartScreen
             internal bool Enabled = true;
             [TomlPrecedingComment("RGBA Color of the Version Text")]
             internal Color TextColor = new Color(1, 1, 1);
-            [TomlPrecedingComment("Toggles the Auto-Alignment of the Version Text")]
+
+            [TomlPrecedingComment("Toggles the Position Auto-Alignment of the Version Text")]
             internal bool AutoAlign = true;
             [TomlPrecedingComment("Custom Position of the Version Text")]
             internal LemonTuple<int, int> CustomPosition = new LemonTuple<int, int>();
+
+            [TomlPrecedingComment("UnityEngine.TextAnchor of the Version Text")]
+            internal TextAnchor Anchor = TextAnchor.MiddleCenter;
+
+
+            [TomlPrecedingComment("If the Loader's name should be colorized.")]
+            internal bool ColorizeLoaderName = true;
+            [TomlPrecedingComment("Version Text to be Displayed.")]
+            internal string Text = "<loaderName/> v<loaderVersion/> Open-Beta";
         }
 
         internal class cProgressText
@@ -114,7 +129,8 @@ namespace MelonLoader.MelonStartScreen
             internal Color InnerColor = new Color(1.00f, 0.23f, 0.42f);
             [TomlPrecedingComment("Outer RGBA Color of the Progress Bar")]
             internal Color OuterColor = new Color(0.47f, 0.97f, 0.39f);
-            [TomlPrecedingComment("Toggles the Auto-Alignment of the Progress Bar")]
+
+            [TomlPrecedingComment("Toggles the Position Auto-Alignment of the Progress Bar")]
             internal bool AutoAlign = true;
             [TomlPrecedingComment("Custom Position of the Progress Bar")]
             internal LemonTuple<int, int> CustomPosition = new LemonTuple<int, int>();